third_party/blink/web_tests/webcodecs/basic_video_encoding.html

<!DOCTYPE html>
<html>

<head>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
</head>

<body>
  <img id='frame_image' style='display: none;' src="pattern.png">
  <script>

    'use strict';

    async function waitTillLoaded(img) {
      if (img.complete)
        return Promise.resolve();
      return new Promise((resolve, reject) => {
        img.onload = () => resolve()
      });
    }

    async function generateBitmap(width, height, text) {
      let img = document.getElementById('frame_image');
      await waitTillLoaded(img);
      let cnv = document.createElement("canvas");
      cnv.height = height;
      cnv.width = width;
      var ctx = cnv.getContext('2d');
      ctx.drawImage(img, 0, 0, width, height);
      ctx.font = '30px fantasy';
      ctx.fillText(text, 5, 40);
      return createImageBitmap(cnv);
    }

    async function createFrame(width, height, ts) {
      let imageBitmap = await generateBitmap(width, height, ts.toString());
      return new VideoFrame(imageBitmap, { timestamp: ts });
    }

    function delay(time_ms) {
      return new Promise((resolve, reject) => {
        setTimeout(resolve, time_ms);
      });
    };

    async function encode_decode_test(codec, acc) {
      const w = 120;
      const h = 60;
      let next_ts = 0
      let frames_to_encode = 16;
      let frames_encoded = 0;
      let frames_decoded = 0;
      let errors = 0;

      let decoder = new VideoDecoder({
        output(frame) {
          assert_equals(frame.cropWidth, w, "cropWidth");
          assert_equals(frame.cropHeight, h, "cropHeight");
          assert_equals(frame.timestamp, next_ts++, "timestamp");
          frames_decoded++;
          frame.destroy();
        },
        error(e) {
          errors++;
          console.log(e.message);
        }
      });

      const encoder_init = {
        output(chunk, config) {
          var data = new Uint8Array(chunk.data);
          if (decoder.state != "configured" || config.description) {
            decoder.configure(config);
          }
          decoder.decode(chunk);
          frames_encoded++;
        },
        error(e) {
          errors++;
          console.log(e.message);
        }
      };

      const encoder_config = {
        codec: codec,
        acceleration: acc,
        width: w,
        height: h,
        bitrate: 5000000,
        framerate: 30,
      };

      let encoder = new VideoEncoder(encoder_init);
      encoder.configure(encoder_config);

      for (let i = 0; i < frames_to_encode; i++) {
        var frame = await createFrame(w, h, i);
        let keyframe = (i % 5 == 0);
        encoder.encode(frame, { keyFrame: keyframe });
        await delay(1);
      }
      await encoder.flush();
      await decoder.flush();
      encoder.close();
      decoder.close();
      assert_equals(frames_encoded, frames_to_encode);
      assert_equals(frames_decoded, frames_to_encode);
      assert_equals(errors, 0);
    }

    async function encode_test(codec, acc) {
      let w = 120;
      let h = 60;
      let next_ts = 0
      let frames_to_encode = 16;
      let frames_processed = 0;
      let errors = 0;

      let process_video_chunk = function (chunk, config) {
        assert_greater_than_equal(chunk.timestamp, next_ts++);
        var data = new Uint8Array(chunk.data);
        let type = (chunk.timestamp % 5 == 0) ? "key" : "delta";
        assert_equals(chunk.type, type);
        assert_greater_than_equal(data.length, 0);
        if (config) {
          assert_equals(config.codec, codec);
          assert_equals(config.codedWidth, w);
          assert_equals(config.codedHeight, h);
          let data = new Uint8Array(config.description);
        }
        frames_processed++;
      };

      const init = {
        output: process_video_chunk,
        error: (e) => {
          error++;
          console.log(e.message);
        },
      };
      const params = {
        codec: codec,
        acceleration: acc,
        width: w,
        height: h,
        bitrate: 5000000,
        framerate: 30,
      };
      let encoder = new VideoEncoder(init);
      encoder.configure(params);
      for (let i = 0; i < frames_to_encode; i++) {
        var frame = await createFrame(w, h, i);
        let keyframe = (i % 5 == 0);
        encoder.encode(frame, { keyFrame: keyframe });
        await delay(1);
      }
      await encoder.flush();
      encoder.close();
      assert_equals(frames_processed, frames_to_encode);
      assert_equals(errors, 0);
    }

    promise_test(encode_test.bind(null, "vp09.00.10.08", "allow"),
      "encoding vp9 profile0");

    promise_test(encode_test.bind(null, "vp09.02.10.10", "allow"),
      "encoding vp9 profile2");

    promise_test(encode_decode_test.bind(null, "vp09.02.10.10", "allow"),
      "encoding and decoding vp9 profile2");

    promise_test(encode_test.bind(null, "vp8", "allow"),
      "encoding vp8");

    promise_test(encode_decode_test.bind(null, "vp8", "allow"),
      "encoding and decoding vp8");

    promise_test(encode_test.bind(null, "avc1.42001E", "allow"),
      "encoding avc1.42001E");

    promise_test(encode_decode_test.bind(null, "avc1.42001E", "allow"),
      "encoding and decoding avc1.42001E");

    /* Uncomment this for manual testing, before we have GPU tests for that */
    // promise_test(encode_test.bind(null, "avc1.42001E", "require"),
    //  "encoding avc1.42001E");

    // promise_test(encode_decode_test.bind(null, "avc1.42001E", "require"),
    //  "encoding and decoding avc1.42001E");


  </script>
</body>

</html>
